// src/app/model/place/place.ts
/**
 * @name:         MET CS 701 - Term Project
 * @author:       Ritesh Gandhi        
 * @description:  Place Model
 */

export class Place {
    customerPosition: any;
    uid: number;
    id: string;
    name: string;
    photoUrl: string;
    open_now: string;
    rating: number;
    address: string;
    types: string[];
    location: any;
}